Ranking to Double

+++ Sports4All/Points.cs
@@ -8,14 +8,14 @@ namespace Sports4All
 {
     static class Points
     {
-        public static readonly int _eventPerformed_Weight = 3;
-        public static readonly int _reservePerformed_Weight = 2;
-        public static readonly int _fairplay_Weight = 4;
-        public static readonly int _skill_Weight = 3;
-        public static readonly int _racio_Weight = 2;
-        public static readonly int _levelChange = 100; // variavel somente usada para determinar a quantidade de pontos necessaria para transitar de nivel
+        public static readonly double _eventPerformed_Weight = 3;
+        public static readonly double _reservePerformed_Weight = 2;
+        public static readonly double _fairplay_Weight = 4;
+        public static readonly double _skill_Weight = 3;
+        public static readonly double _racio_Weight = 2;
+        public static readonly double _levelChange = 100; // variavel somente usada para determinar a quantidade de pontos necessaria para transitar de nivel
 
-        public static readonly int _ParkPrice_Weight = 3;
-        public static readonly int _ParkQuality_Weight = 4;
+        public static readonly double _ParkPrice_Weight = 3;
+        public static readonly double _ParkQuality_Weight = 4;
     }
 }
